BACKEND/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "pymongo[srv]>=4.17.0",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
